% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_euiss.R
\name{annotate_euiss}
\alias{annotate_euiss}
\title{Custom `ggplot2` annotation}
\usage{
annotate_euiss(
  geom = "text",
  lineheight = 0.85,
  family = txt_family,
  size = txt_label,
  hjust = "left",
  ...
)
}
\arguments{
\item{geom}{`ggplot2` geom for annotation}

\item{lineheight}{text line height}

\item{family}{text family}

\item{size}{text size}

\item{hjust}{text alignment}

\item{...}{other arguments that can be passed to `ggplot2::annotate()`}
}
\value{
Pre-styled `annotate()`
}
\description{
Custom `ggplot2` annotation
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt)) +
  geom_point_euiss() +
  annotate_euiss(label = "annotation in center-top", x = 300, y = 5)
}
