% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_col_horizontal.R
\name{chart_col_horizontal}
\alias{chart_col_horizontal}
\title{EUISS column chart}
\usage{
chart_col_horizontal(
  data,
  x,
  y,
  label,
  title = "",
  subtitle = "",
  caption = "",
  nudge_y = 0.5,
  vjust = 0,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in the plot.}

\item{x}{x-aesthetic passed into each layer.}

\item{y}{y-aesthetic passed into each layer.}

\item{label}{label for maximum x-value}

\item{title}{chart title}

\item{subtitle}{chart subtitle}

\item{caption}{chart caption}

\item{nudge_y}{Vertical adjustment to nudge labels by.}

\item{vjust}{Vertical alignment; defaults to 0 (bottom).}

\item{...}{Other arguments passed on to layer().}
}
\value{
A complete horizontal column chart with labels and titles, that can be amended with further \code{ggplot2} geoms.
}
\description{
EUISS column chart
}
\examples{
require(ggplot2)
require(magrittr)
mtcars \%>\% 
  chart_col_horizontal(x = mpg, y = cyl)
}
