% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_line.R
\name{chart_line}
\alias{chart_line}
\title{EUISS line chart}
\usage{
chart_line(
  data,
  x,
  y,
  group,
  label,
  title = "",
  subtitle = "",
  caption = "",
  ...
)
}
\arguments{
\item{data}{The data to be displayed in the plot.}

\item{x}{x-aesthetic passed into each layer.}

\item{y}{y-aesthetic passed into each layer.}

\item{group}{group-aesthetic for line(s).}

\item{label}{label for maximum x-value}

\item{title}{chart title}

\item{subtitle}{chart subtitle}

\item{caption}{chart caption}

\item{...}{Other arguments passed on to layer().}
}
\value{
A complete line chart with labels and titles, that can be amended with further \code{ggplot2} geoms.
}
\description{
EUISS line chart
}
\examples{
require(magrittr)
mtcars \%>\% 
  chart_line(x = mpg, y = drat, group = cyl)
}
