% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_utilities.R
\name{euiss_color_continuous}
\alias{euiss_color_continuous}
\title{Enhanced continuous color scale using EUISS palette}
\usage{
euiss_color_continuous()
}
\description{
Enhanced continuous color scale using EUISS palette
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = cyl)) +
  ggplot2::geom_point() +
  euiss_color_continuous()
}
