% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_utilities.R
\name{euiss_color_discrete}
\alias{euiss_color_discrete}
\title{Enhanced discrete color scale using EUISS palette}
\usage{
euiss_color_discrete()
}
\description{
Enhanced discrete color scale using EUISS palette
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = base::factor(cyl))) +
  ggplot2::geom_point() +
  euiss_color_discrete()
}
