% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_utilities.R
\name{euiss_force_import_fonts}
\alias{euiss_force_import_fonts}
\title{Force import/register fonts found in system but not in extrafont}
\usage{
euiss_force_import_fonts()
}
\description{
Force import/register fonts found in system but not in extrafont
}
\examples{
\dontrun{
euiss_force_import_fonts()
}
}
