% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_utilities.R
\name{euiss_mar0}
\alias{euiss_mar0}
\title{Create margin theme with minimal margins (like mar0 in style script)}
\usage{
euiss_mar0(t = 0.1, r = 0.1, b = 0.1, l = 0.1)
}
\arguments{
\item{t}{top margin in mm}

\item{r}{right margin in mm}

\item{b}{bottom margin in mm}

\item{l}{left margin in mm}
}
\description{
Create margin theme with minimal margins (like mar0 in style script)
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + 
  geom_point() + 
  euiss_mar0()
}
