% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{euiss_set_fonts}
\alias{euiss_set_fonts}
\title{Set fonts for eiussR package}
\usage{
euiss_set_fonts(family = NULL, bold = NULL, height = NULL, label_size = NULL)
}
\arguments{
\item{family}{Font family name}

\item{bold}{Font weight for bold text}

\item{height}{Line height}

\item{label_size}{Text label size}
}
\description{
Set fonts for eiussR package
}
