% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_colors.R
\name{euiss_show_colors}
\alias{euiss_show_colors}
\title{Display EUISS color palette}
\usage{
euiss_show_colors(plot = FALSE, category = "all")
}
\arguments{
\item{plot}{Logical, whether to create a visual plot}

\item{category}{Character: "all", "basic", "cold", "warm", "green"}
}
\value{
Invisibly returns color list
}
\description{
Display EUISS color palette
}
\examples{
# Print colors
euiss_show_colors()

# Visual plot
euiss_show_colors(plot = TRUE)
euiss_show_colors(plot = TRUE, category = "warm")
}
