% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euissR.R
\docType{package}
\name{euissR-package}
\alias{euissR}
\alias{euissR-package}
\title{\code{euissR} package}
\description{
Custom ggplot2 theme, geoms and color palettes to reflect the visual identity of the EU Institute for Security Studies.
}
\details{
See the README on
\href{https://github.com/cdtrich/euissR/blob/main/README.md}{GitHub}
}
\keyword{internal}
