% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_area_euiss.R
\name{geom_area_euiss}
\alias{geom_area_euiss}
\title{Custom ggplot2::geom_area()}
\usage{
geom_area_euiss(
  fill = scales::alpha(teal, 0.1),
  col = teal,
  outline.type = "upper",
  ...
)
}
\arguments{
\item{fill}{area color}

\item{col}{line color}

\item{outline.type}{line position}

\item{...}{other arguments that can be passed to `ggplot2::geom_area()`}
}
\value{
Pre-styled `geom_area()`
}
\description{
Custom ggplot2::geom_area()
}
\examples{
require(dplyr)
mtcars \%>\% 
  dplyr::group_by(gear, cyl) \%>\% 
  dplyr::summarize(wt = sum(wt)) \%>\% 
  ggplot2::ggplot(ggplot2::aes(cyl, wt, group = gear)) + 
  geom_area_euiss() + 
  ggplot2::facet_wrap(~gear)
}
