% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_euiss.R
\name{geom_bar_euiss}
\alias{geom_bar_euiss}
\title{Custom ggplot2::geom_bar()}
\usage{
geom_bar_euiss(width = 0.25, ...)
}
\arguments{
\item{width}{bar width}

\item{...}{other arguments that can be passed to `ggplot2::geom_bar()`}
}
\value{
Pre-styled `geom_bar()`
}
\description{
Custom ggplot2::geom_bar()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(cyl, fill = base::factor(gear))) + 
  geom_bar_euiss()
}
