% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_col_euiss.R
\name{geom_col_euiss}
\alias{geom_col_euiss}
\title{Custom ggplot2::geom_col()}
\usage{
geom_col_euiss(width = 0.25, ...)
}
\arguments{
\item{width}{column width}

\item{...}{other arguments that can be passed to `ggplot2::geom_col()`}
}
\value{
Pre-styled `geom_col()`
}
\description{
Custom ggplot2::geom_col()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(cyl, hp, fill = base::factor(gear))) + 
  geom_col_euiss()
}
