% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line_euiss.R
\name{geom_line_euiss}
\alias{geom_line_euiss}
\title{Custom ggplot2::geom_line()}
\usage{
geom_line_euiss(col = teal, size = lwd_line, ...)
}
\arguments{
\item{col}{line color}

\item{size}{line width}

\item{...}{other arguments that can be passed to `ggplot2::geom_line()`}
}
\value{
Pre-styled `geom_line()`
}
\description{
Custom ggplot2::geom_line()
}
\examples{
require(dplyr)
mtcars \%>\% 
  dplyr::group_by(gear, cyl) \%>\% 
  dplyr::summarize(wt = sum(wt)) \%>\% 
  ggplot2::ggplot(ggplot2::aes(cyl, wt, group = gear)) + 
  geom_line_euiss() + 
  ggplot2::facet_wrap(~gear)
}
