% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_linerange_x_euiss.R
\name{geom_linerange_x_euiss}
\alias{geom_linerange_x_euiss}
\title{Custom horizontal ggplot2::geom_linerange()}
\usage{
geom_linerange_x_euiss(col = teal, size = lwd_line, ...)
}
\arguments{
\item{col}{line color}

\item{size}{line width}

\item{...}{other arguments that can be passed to `ggplot2::geom_linerange()`}
}
\value{
Pre-styled `geom_linerange()`
}
\description{
Custom horizontal ggplot2::geom_linerange()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = hp, y = wt, xmax = hp)) + 
  geom_linerange_x_euiss() + 
  geom_point_euiss()
}
