% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_euiss.R
\name{geom_point_euiss}
\alias{geom_point_euiss}
\title{Custom ggplot2::geom_point()}
\usage{
geom_point_euiss(col = teal, fill = NA, size = 2, shape = 21, ...)
}
\arguments{
\item{col}{point color}

\item{fill}{point fill}

\item{size}{point size}

\item{shape}{any shape value that can be passed to `ggplot2::geom_point()`}

\item{...}{other arguments that can be passed to `ggplot2::geom_point()`}
}
\value{
Pre-styled `geom_point()`
}
\description{
Custom ggplot2::geom_point()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt)) + 
  geom_point_euiss()
}
