% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{lwd_grid}
\alias{lwd_grid}
\title{width and height measurements for EUISS publications}
\format{
measurements (in mm) are stored in tibbles per type of publication.
}
\usage{
lwd_grid
}
\description{
width and height measurements for EUISS publications
}
\keyword{datasets}
