% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_enhanced_euiss.R
\name{pal_cat_euiss_enhanced}
\alias{pal_cat_euiss_enhanced}
\title{Enhanced euissR palettes with additional variants}
\usage{
pal_cat_euiss_enhanced(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
vector of n colors
}
\description{
Enhanced categorical palette using updated colors from style script
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = factor(cyl))) + 
  ggplot2::geom_point() + 
  ggplot2::scale_color_manual(values = pal_cat_euiss_enhanced(6))
}
