% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_enhanced_euiss.R
\name{pal_seq_bathy_euiss}
\alias{pal_seq_bathy_euiss}
\title{Bathymetry palette for depth visualization}
\usage{
pal_seq_bathy_euiss(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
vector of n colors
}
\description{
Bathymetry palette for depth visualization
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = disp)) + 
  ggplot2::geom_point() + 
  ggplot2::scale_color_gradientn(colors = pal_seq_bathy_euiss(6))
}
