% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euiss_mar0.R
\name{theme_euiss_mar0}
\alias{theme_euiss_mar0}
\title{euissR theme with narrow margins}
\usage{
theme_euiss_mar0(t = 0.1, r = 0.1, b = 0.1, l = 0.1, unit = "mm", ...)
}
\arguments{
\item{t}{top margin}

\item{r}{right margin}

\item{b}{bottom margin}

\item{l}{left margin}

\item{unit}{unit of margin values}

\item{...}{other arguments that can be passed to `ggplot2::margin()`}
}
\value{
ggplot2 `theme` object
}
\description{
euissR theme with narrow margins
}
