% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{width_cp_onecol}
\alias{width_cp_onecol}
\title{euissR measures}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
width_cp_onecol
}
\description{
euissR measures
}
\keyword{datasets}
