% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_axis_euiss.R
\name{x_axis_euiss}
\alias{x_axis_euiss}
\title{Custom ggplot2::geom_hline()}
\usage{
x_axis_euiss(yintercept = 0, size = lwd_grid, col = col_grid, ...)
}
\arguments{
\item{yintercept}{position of x-axis. `default = 0`}

\item{size}{line width}

\item{col}{line color}

\item{...}{other arguments that can be passed to `ggplot2::geom_hline()`}
}
\value{
Pre-styled `geom_hline()` that inserts a grey x-axis at y-intercept = 0
}
\description{
Custom ggplot2::geom_hline()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt)) + 
  geom_point_euiss() + 
  x_axis_euiss()
}
