% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.copy_colors_to_namespace}
\alias{.copy_colors_to_namespace}
\title{Copy colors from .euiss_env to package namespace}
\usage{
.copy_colors_to_namespace()
}
\description{
Copy colors from .euiss_env to package namespace
}
\keyword{internal}
