% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line_euiss.R
\name{geom_line_euiss}
\alias{geom_line_euiss}
\title{Custom ggplot2::geom_line()}
\usage{
geom_line_euiss(
  mapping = NULL,
  data = NULL,
  col = teal,
  linewidth = lwd_line,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{col}{line color}

\item{show.legend}{logical. Should this layer be included in the legends? \code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. \code{borders()}.}

\item{...}{other arguments that can be passed to \code{ggplot2::geom_line()}}

\item{size}{line width}
}
\value{
Pre-styled \code{geom_line()}
}
\description{
Custom ggplot2::geom_line()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt, label = wt)) + 
  geom_line_euiss()
}
