% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_euiss.R
\name{geom_text_euiss}
\alias{geom_text_euiss}
\title{Custom ggplot2::geom_text()}
\usage{
geom_text_euiss(
  lineheight = 0.85,
  family = txt_family,
  size = txt_label,
  hjust = "left",
  ...
)
}
\arguments{
\item{lineheight}{text line height}

\item{family}{text family}

\item{size}{text size}

\item{hjust}{text alignment}

\item{...}{other arguments that can be passed to `ggplot2::geom_text()`}
}
\value{
Pre-styled `geom_text()`
}
\description{
Custom ggplot2::geom_text()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt, label = wt)) + 
  geom_text_euiss()
}
