% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_euiss.R
\name{ggsave_euiss}
\alias{ggsave_euiss}
\title{Custom ggplot2::ggsave()}
\usage{
ggsave_euiss(
  filename,
  plot = ggplot2::last_plot(),
  publication = c("book", "cp", "brief"),
  w = NA,
  h = 1,
  units = "mm",
  dev = NULL,
  ...
)
}
\arguments{
\item{filename}{name of file, incl. sub-folders, *incl. file extension*}

\item{plot}{last plot}

\item{publication}{\code{character}, one of \code{book}, \code{brief} or \code{cp}}

\item{w}{\code{character}, one of \code{onecol}, \code{twocol} or \code{full} for width}

\item{h}{\code{numeric} fraction of full height}

\item{units}{units of width and height, defaults to \code{mm}}

\item{dev}{optional explicit device (e.g. \code{svglite::svglite}); otherwise inferred}

\item{...}{other arguments that can be passed to \code{ggplot2::ggsave()}}
}
\value{
Parametrized \code{ggsave()}
}
\description{
Custom ggplot2::ggsave()
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(disp, wt, label = wt)) + 
  geom_text_euiss() +
  geom_line_euiss() 
  # NOT RUN 
  # ggsave_euiss("test.pdf", publication = "cp", w = "onecol", h = 1)
  # END NOT RUN 
}
