% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_div_euiss.R
\name{pal_div_euiss}
\alias{pal_div_euiss}
\title{euissR palettes based on `euissR::colors`}
\usage{
pal_div_euiss(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
euissR palettes based on `euissR::colors`
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = base::factor(cyl))) + 
  ggplot2::geom_point() + 
  ggplot2::scale_color_manual(values = pal_div_euiss(6))
}
