% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_seq_r_light_euiss.R
\name{pal_seq_r_light_euiss}
\alias{pal_seq_r_light_euiss}
\title{euissR palettes based on `euissR::colors`}
\usage{
pal_seq_r_light_euiss(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
vector of n colors
}
\description{
euissR palettes based on `euissR::colors`
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = cyl)) + 
  ggplot2::geom_point() + 
  ggplot2::scale_color_gradientn(colors = pal_seq_r_light_euiss(4))
}
