% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euiss_utilities.R
\name{scale_color_continuous_euiss}
\alias{scale_color_continuous_euiss}
\title{Enhanced continuous color scale using EUISS palette}
\usage{
scale_color_continuous_euiss()
}
\description{
Enhanced continuous color scale using EUISS palette
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = cyl)) +
  ggplot2::geom_point() +
  scale_color_continuous_euiss()
}
