% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_discrete_euiss.R
\name{scale_color_discrete_euiss}
\alias{scale_color_discrete_euiss}
\title{euissR scales based on `euissR::scales`}
\usage{
scale_color_discrete_euiss()
}
\description{
euissR scales based on `euissR::scales`
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(wt, mpg, col = base::factor(cyl))) +
  ggplot2::geom_point() +
  scale_color_discrete_euiss()
}
