% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euiss.R
\name{theme_euiss}
\alias{theme_euiss}
\title{Custom \code{ggplot2} theme}
\usage{
theme_euiss()
}
\description{
Custom \code{ggplot2} theme
}
\details{
\code{theme_euiss()} runs when \code{euissR} package is loaded.
}
